% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chain}
\alias{chain}
\title{Synthetic data representing a simple mediator chain}
\format{
A data set with six columns, each is a factor of two levels.
}
\usage{
data(chain)
}
\description{
Chain is generated from a simple Bayes network,
\deqn{X\rightarrow M_1 \rightarrow M_2  \rightarrow M_3  \rightarrow M_4  \rightarrow Y}
where every variable is binary.
The set consists of 11 observations, and is tuned to be easily deciphered.
}
\keyword{datasets}
