% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Extract the vertex hierarchy from the vistla tree}
\usage{
hierarchy(x)
}
\arguments{
\item{x}{vistla object.}
}
\value{
A data frame of a class \code{vistla_hierarchy}.
}
\description{
Traverses the vistla tree in a depth-first order and 
lists the visited vertices as a data frame.
}
\note{
This function effectively prunes the tree off suboptimal paths.
}
