% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junction}
\alias{junction}
\title{Synthetic data representing a junction}
\format{
A data set with eight factor columns.
}
\usage{
data(junction)
}
\description{
Junction is a model of a multimodal agent, a variable that is an element of multiple separate paths.
Here, these paths are 
\eqn{Y\rightarrow A_1\rightarrow A_2\rightarrow J \rightarrow A_3}
and 
\eqn{Y\rightarrow B_1\rightarrow B_2\rightarrow J \rightarrow B_3,}
while \eqn{J} is the junction.
The set consists of 50 observations.
}
\keyword{datasets}
