% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistla.R
\name{leaf_scores}
\alias{leaf_scores}
\title{Extract leaf scores of vertex pairs}
\usage{
leaf_scores(x)
}
\arguments{
\item{x}{vistla object.}
}
\value{
A square matrix with leaf scores of all feature pairs.
}
\description{
Produces a matrix \eqn{S} where \eqn{S_{ij}} is a score
 of the path ending in vertices \eqn{i} and \eqn{j}.
Since vistla works on vertex pairs, this value is unique.
This can be interpreted as a feature similarity matrix
 in context of the current vistla root.
}
\note{
This function should be called on an unpruned vistla tree,
 otherwise the result will be mostly composed of zeroes.
}
