% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistla.R
\name{mi_scores}
\alias{mi_scores}
\title{Extract mutual information score matrix}
\usage{
mi_scores(x)
}
\arguments{
\item{x}{vistla object.}
}
\value{
A symmetric square matrix with mutual information scores between features and root.
}
\description{
Produces a matrix \eqn{S} where \eqn{S_{ij}} is a 
 value of \eqn{I(X_i;X_j)}.
This matrix is always calculated as an initial step of the
 vistla algorithm and stored in the vistla object.
}
