% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_to}
\alias{path_to}
\title{Extract a single path}
\usage{
path_to(x, target, detailed = FALSE)
}
\arguments{
\item{x}{vistla object.}

\item{target}{target feature name.}

\item{detailed}{if \code{TRUE}, suppresses default output and presents the same paths as a data frame featuring score.}
}
\value{
By default, a character vector with names of features along the path from \code{target} into root.
 When \code{detailed} is set to \code{TRUE}, a \code{data.frame} in a format identical to this produced by
 \code{\link{branches}}, yet without the \code{leaf} column.
}
\description{
Gives access to a vector of feature names over a path to a certain target feature.
}
