% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{List all paths}
\usage{
paths(x, targets_only = !is.null(x$targets), detailed = FALSE)
}
\arguments{
\item{x}{vistla object.}

\item{targets_only}{if \code{TRUE}, only paths to targets are extracted. 
By default, turned on when \code{x} has targets, and off otherwise.}

\item{detailed}{passed to \code{\link{path_to}}. If \code{TRUE},
suppresses default output and presents the same paths in a form of
data frames featuring score.}
}
\value{
A named list with one element per leaf or target, containing
 the path between this feature and root, in a format identical
 to this used by the \code{\link{path_to}} function.
}
\description{
Executes \code{\link{path_to}} for all path possible targets and returns
a list with the results.
}
