% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/vistla.R
\name{print.vistla_hierarchy}
\alias{print.vistla_hierarchy}
\alias{print.vistla}
\title{Print vistla objects}
\usage{
\method{print}{vistla_hierarchy}(x, ...)

\method{print}{vistla}(x, n = 7L, ...)
}
\arguments{
\item{x}{vistla object.}

\item{...}{ignored.}

\item{n}{maximal number of paths to preview.}
}
\value{
Invisible copy of \code{x}.
}
\description{
Utility functions to print vistla objects.
}
