% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistla.R
\name{vistla_coerce}
\alias{vistla_coerce}
\title{Coerce data as vistla would}
\usage{
vistla_coerce(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
\code{x} coerced into factor.
}
\description{
This function will coerce the input vector into factor as \code{\link{vistla}} function would.
Useful for testing or pre-computing quantisation.
}
