% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junction}
\alias{junction}
\title{Synthetic data representing a junction}
\format{
A data set with five factor columns.
}
\usage{
data(junction)
}
\description{
Junction is a model of a multimodal agent, a variable that is an element of multiple separate paths.
Here, these paths are \eqn{A_1\rightarrow J \rightarrow A_2} and 
\eqn{B_1\rightarrow J \rightarrow B_2,}
while \eqn{J} is the junction.
The set consists of 12 observations, and is tuned to be easily deciphered.
}
\keyword{datasets}
