% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vistla}
\alias{plot.vistla}
\title{Overview plot of the vistla tree}
\usage{
\method{plot}{vistla}(x, ..., scale_width = TRUE)
}
\arguments{
\item{x}{vistla object.}

\item{...}{additional graphical parameters, passed to plot.}

\item{scale_width}{if \code{TRUE}, widths of links are scaled according to score.}
}
\value{
\code{x}, invisibly.
}
\description{
Overview plot of the vistla tree
}
